requirements.txt
transformers>=4.0
torch>=1.0
matplotlib>=3.0
pandas>=1.0
Django~=4.2.16
paypal~=1.2.5
diffusers~=0.31.0
django-recaptcha~=4.0.0
djangorestframework~=3.15.2
djangorestframework-simplejwt~=5.3.1
pillow~=11.0.0